create job presenter
